Fix cpplint errors in cc/(animation|input|layers|trees|test)/
This fixes lint errors in everything except for the few stragglers (occlusion unit test, etc) that haven't been chromified.

BUG=144577


Review URL: https://chromiumcodereview.appspot.com/12965007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@190767 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bf691c24ca3cf0ebbe8c82e871b4b05bd4256f6d

+++ cc/scheduler/delay_based_time_source_unittest.cc
@@ -4,6 +4,7 @@
 
 #include "cc/scheduler/delay_based_time_source.h"
 
+#include "base/basictypes.h"
 #include "cc/base/thread.h"
 #include "cc/test/scheduler_test_common.h"
 #include "testing/gtest/include/gtest/gtest.h"
@@ -303,7 +304,7 @@ TEST(DelayBasedTimeSourceTest, AchievesTargetRateWithNoNoise) {
 
   double total_frame_time = 0.0;
   for (int i = 0; i < num_iterations; ++i) {
-    long long delay_ms = thread.PendingDelayMs();
+    int64 delay_ms = thread.PendingDelayMs();
 
     // accumulate the "delay"
     total_frame_time += delay_ms / 1000.0;
